introduce different folder structure for introspection tests, more notes on refactoring and additional needed tests
