+++ introspection-engine/connectors/sql-introspection-connector/tests/db_specific_introspection/mysql/mod.rs
@@ -1,0 +1,3 @@
+pub mod relations_mysql;
+pub mod remapping_database_names_mysql;
+pub mod tables_mysql;
