+++ introspection-engine/connectors/sql-introspection-connector/tests/db_specific_introspection/postgres/mod.rs
@@ -1,0 +1,5 @@
+pub mod native_lists_postgres;
+pub mod relations_postgres;
+pub mod relations_with_compound_fk_postgres;
+pub mod remapping_database_names_postgres;
+pub mod tables_postgres;
