+++ introspection-engine/connectors/sql-introspection-connector/tests/db_specific_introspection/sqlite/mod.rs
@@ -1,0 +1,3 @@
+pub mod relations_sqlite;
+pub mod remapping_database_names_sqlite;
+pub mod tables_sqlite;
