+++ introspection-engine/connectors/sql-introspection-connector/tests/rpc_calls/mod.rs
@@ -1,9 +1,3 @@
-#![allow(dead_code)]
-
 mod get_metadata_command_for_empty_db_tests;
 mod get_metadata_command_tests;
 mod list_databases_command_tests;
-mod mysql;
-mod postgres;
-mod remapping_invalid_characters;
-mod sqlite;
